refactor(platform/rest): Update REST API to use standard FastAPI structure (#8519)
